fix:CIのfail修正とルーティング変更 - CIのFail原因を修正 - 画面起動時のURLを修正
