+++ src/ui/src/components/Sidebar.tsx
@@ -18,7 +18,7 @@ export default function Sidebar({ isOpen, onClose, onToggle, isMobile = false }:
 
   const menuItems = [
     {
-      href: '/',
+      href: '/upload',
       icon: (
         <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
           <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12" />
@@ -93,7 +93,8 @@ export default function Sidebar({ isOpen, onClose, onToggle, isMobile = false }:
         <nav className="flex-1 mt-8 px-4 overflow-y-auto" data-testid="sidebar-navigation">
           <ul className="space-y-2">
             {menuItems.map((item) => {
-              const isActive = pathname === item.href;
+              // ルート / の場合は /upload メニューをアクティブにする
+              const isActive = pathname === item.href || (pathname === '/' && item.href === '/upload');
               return (
                 <li key={item.href}>
                   <Link
